+++ src/Zombie.java
@@ -6,7 +6,7 @@ public class Zombie extends Character {
     float movementSpeed;
     float damage;
 
-    public Zombie(float health, int lane, int[] position, float movementSpeed, float damage, ImageView image) {
+    public Zombie(float health, int lane, double[] position, float movementSpeed, float damage, ImageView image) {
         super(health, lane, position,image);
         this.movementSpeed=movementSpeed;
         this.damage=damage;
@@ -14,22 +14,22 @@ public class Zombie extends Character {
 }
 
 class BucketZombie extends  Zombie{
-    public BucketZombie(int lane, int[] position,ImageView image) {
+    public BucketZombie(int lane, double[] position,ImageView image) {
         super(300, lane, position,(float) 1.5,10,image);
     }
 }
 class ConeZombie extends Zombie{
-    public ConeZombie(float health, int lane, int[] position,ImageView image) {
+    public ConeZombie(float health, int lane, double[] position,ImageView image) {
         super(250, lane, position,1,15,image);
     }
 }
 class BasicZombie extends  Zombie{
-    public BasicZombie(float health, int lane, int[] position,ImageView image) {
+    public BasicZombie(float health, int lane, double[] position,ImageView image) {
         super(100, lane, position,1,20,image);
     }
 }
 class JavelinZombie extends Zombie{
-    public JavelinZombie(float health, int lane, int[] position,ImageView image) {
+    public JavelinZombie(float health, int lane, double[] position,ImageView image) {
         super(150, lane, position,1,10,image);
     }
 }
